refactor(console): update conversion report timing
