+++ packages/console/src/components/Conversion/utils.ts
@@ -1,6 +1,7 @@
 import { cond } from '@silverhand/essentials';
+import debug from 'debug';
 
-import { isProduction } from '@/consts/env';
+const log = debug('conversion');
 
 export const gtagAwTrackingId = 'AW-11124811245';
 export enum GtagConversionId {
@@ -54,30 +55,18 @@ export const hashEmail = async (email?: string) => {
   return sha256(canonicalizedEmail);
 };
 
-/** Print debug message if not in production. */
-const debug = (...args: Parameters<(typeof console)['debug']>) => {
-  if (!isProduction) {
-    console.debug(...args);
-  }
-};
-
 /**
  * Add more if needed: https://reddit.my.site.com/helpcenter/s/article/Install-the-Reddit-Pixel-on-your-website
  */
-export type RedditReportType =
-  | 'PageVisit'
-  | 'ViewContent'
-  | 'Search'
-  | 'Purchase'
-  | 'Lead'
-  | 'SignUp';
+type RedditReportType = 'PageVisit' | 'ViewContent' | 'Search' | 'Purchase' | 'Lead' | 'SignUp';
 
-export const reportToReddit = (redditType: RedditReportType) => {
+const reportToReddit = (redditType: RedditReportType) => {
   if (!window.rdt) {
+    log('report:', 'window.rdt is not available');
     return false;
   }
 
-  debug('report:', 'redditType =', redditType);
+  log('report:', 'redditType =', redditType);
   window.rdt('track', redditType);
 
   return true;
@@ -88,13 +77,14 @@ export const reportToGoogle = (
   { transactionId }: { transactionId?: string } = {}
 ) => {
   if (!window.gtag) {
+    log('report:', 'window.gtag is not available');
     return false;
   }
 
   const run = async () => {
     const transaction = cond(transactionId && { transaction_id: await sha256(transactionId) });
 
-    debug('report:', 'gtagId =', gtagId, 'transaction =', transaction);
+    log('report:', 'gtagId =', gtagId, 'transaction =', transaction);
     window.gtag?.('event', 'conversion', {
       send_to: gtagId,
       ...transaction,
@@ -104,4 +94,26 @@ export const reportToGoogle = (
   void run();
 
   return true;
+};
+
+type ReportConversionOptions = {
+  transactionId?: string;
+  gtagId?: GtagConversionId;
+  redditType?: RedditReportType;
+};
+
+export const reportConversion = async ({
+  gtagId,
+  redditType,
+  transactionId,
+}: ReportConversionOptions) => {
+  if (!shouldReport) {
+    log('skip reporting conversion:', { gtagId, redditType, transactionId });
+    return;
+  }
+
+  return Promise.all([
+    gtagId ? reportToGoogle(gtagId, { transactionId }) : undefined,
+    redditType ? reportToReddit(redditType) : undefined,
+  ]);
 };
